Removing converter prefix name
