+++ vraptor-core/src/main/java/br/com/caelum/vraptor/converter/PrimitiveFloatConverter.java
@@ -14,11 +14,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package br.com.caelum.vraptor.converter;
 
 import static com.google.common.base.Strings.isNullOrEmpty;
 
-import java.math.BigDecimal;
 import java.text.DecimalFormat;
 import java.text.NumberFormat;
 import java.text.ParseException;
@@ -30,46 +30,45 @@ import javax.inject.Inject;
 import br.com.caelum.vraptor.Convert;
 
 /**
- * Localized version of VRaptor's BigDecimal converter. If the input value if empty or a null string, null value is 
- * returned. If the input string is not a number a {@link ConversionException} will be throw.
+ * Localized version of VRaptor's Float converter. If the input value if empty or a null string, null value is returned. 
+ * If the input string is not a number a {@link ConversionException} will be throw.
  *
+ * @author Rafael Dipold
  * @author Otávio Scherer Garcia
- * @since 3.1.2
+ * @since 3.4
  */
-@Convert(BigDecimal.class)
+@Convert(float.class)
 @RequestScoped
-public class LocaleBasedBigDecimalConverter implements Converter<BigDecimal> {
+public class PrimitiveFloatConverter implements Converter<Float> {
 
 	private final Locale locale;
 
 	/** 
 	 * @deprecated CDI eyes only
 	 */
-	protected LocaleBasedBigDecimalConverter() {
+	protected PrimitiveFloatConverter() {
 		this(null);
 	}
 
 	@Inject
-	public LocaleBasedBigDecimalConverter(Locale locale) {
+	public PrimitiveFloatConverter(Locale locale) {
 		this.locale = locale;
 	}
 
 	@Override
-	public BigDecimal convert(String value, Class<? extends BigDecimal> type) {
+	public Float convert(String value, Class<? extends Float> type) {
 		if (isNullOrEmpty(value)) {
-			return null;
+			return 0f;
 		}
 
 		try {
-			return (BigDecimal) getNumberFormat().parse(value);
+			return getNumberFormat().parse(value).floatValue();
 		} catch (ParseException e) {
 			throw new ConversionException(new ConversionMessage("is_not_a_valid_number", value));
 		}
 	}
 	
 	protected NumberFormat getNumberFormat() {
-		DecimalFormat fmt = ((DecimalFormat) DecimalFormat.getInstance(locale));
-		fmt.setParseBigDecimal(true);
-		return fmt;
+		return DecimalFormat.getInstance(locale);
 	}
 }
