+++ vraptor-core/src/test/java/br/com/caelum/vraptor/converter/BigDecimalConverterTest.java
@@ -25,6 +25,7 @@ import static org.hamcrest.Matchers.nullValue;
 import static org.junit.Assert.fail;
 import static org.mockito.Mockito.when;
 
+import java.math.BigDecimal;
 import java.util.Locale;
 
 import javax.servlet.ServletContext;
@@ -36,14 +37,14 @@ import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 
 import br.com.caelum.vraptor.converter.ConversionException;
-import br.com.caelum.vraptor.converter.LocaleBasedDoubleConverter;
+import br.com.caelum.vraptor.converter.BigDecimalConverter;
 import br.com.caelum.vraptor.http.MutableRequest;
 
-public class LocaleBasedDoubleConverterTest {
+public class BigDecimalConverterTest {
 
 	static final String LOCALE_KEY = "javax.servlet.jsp.jstl.fmt.locale";
 
-	private LocaleBasedDoubleConverter converter;
+	private BigDecimalConverter converter;
 	private @Mock MutableRequest request;
 	private @Mock HttpSession session;
 	private @Mock ServletContext context;
@@ -54,36 +55,36 @@ public class LocaleBasedDoubleConverterTest {
 
 		when(request.getServletContext()).thenReturn(context);
 
-		converter = new LocaleBasedDoubleConverter(new Locale("pt", "BR"));
+		converter = new BigDecimalConverter(new Locale("pt", "BR"));
 	}
 
 	@Test
 	public void shouldBeAbleToConvertWithPTBR() {
-		assertThat(converter.convert("10,00", Double.class), is(equalTo(new Double("10.00"))));
-		assertThat(converter.convert("10,01", Double.class), is(equalTo(new Double("10.01"))));
+		assertThat(converter.convert("10,00", BigDecimal.class), is(equalTo(new BigDecimal("10.00"))));
+		assertThat(converter.convert("10,01", BigDecimal.class), is(equalTo(new BigDecimal("10.01"))));
 	}
 
 	@Test
 	public void shouldBeAbleToConvertWithENUS() {
-		converter = new LocaleBasedDoubleConverter(new Locale("en", "US"));
-		assertThat(converter.convert("10.00", Double.class), is(equalTo(new Double("10.00"))));
-		assertThat(converter.convert("10.01", Double.class), is(equalTo(new Double("10.01"))));
+		converter = new BigDecimalConverter(new Locale("en", "US"));
+		assertThat(converter.convert("10.00", BigDecimal.class), is(equalTo(new BigDecimal("10.00"))));
+		assertThat(converter.convert("10.01", BigDecimal.class), is(equalTo(new BigDecimal("10.01"))));
 	}
 
 	 @Test
 	 public void shouldBeAbleToConvertEmpty() {
-		 assertThat(converter.convert("", Double.class), is(nullValue()));
+		 assertThat(converter.convert("", BigDecimal.class), is(nullValue()));
 	 }
 
 	 @Test
 	 public void shouldBeAbleToConvertNull() {
-		 assertThat(converter.convert(null, Double.class), is(nullValue()));
+		 assertThat(converter.convert(null, BigDecimal.class), is(nullValue()));
 	 }
 
 	@Test
 	public void shouldThrowExceptionWhenUnableToParse() {
 		try {
-			converter.convert("vr3.9", Double.class);
+			converter.convert("vr3.9", BigDecimal.class);
 			fail("Should throw exception");
 		} catch (ConversionException e) {
 			assertThat(e.getValidationMessage(), hasMessage("vr3.9 is not a valid number."));
