+++ vraptor-core/src/test/java/br/com/caelum/vraptor/converter/PrimitiveDoubleConverterTest.java
@@ -21,7 +21,6 @@ import static br.com.caelum.vraptor.VRaptorMatchers.hasMessage;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
-import static org.hamcrest.Matchers.nullValue;
 import static org.junit.Assert.fail;
 import static org.mockito.Mockito.when;
 
@@ -36,14 +35,14 @@ import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 
 import br.com.caelum.vraptor.converter.ConversionException;
-import br.com.caelum.vraptor.converter.LocaleBasedFloatConverter;
+import br.com.caelum.vraptor.converter.PrimitiveDoubleConverter;
 import br.com.caelum.vraptor.http.MutableRequest;
 
-public class LocaleBasedFloatConverterTest {
+public class PrimitiveDoubleConverterTest {
 
 	static final String LOCALE_KEY = "javax.servlet.jsp.jstl.fmt.locale";
 
-	private LocaleBasedFloatConverter converter;
+	private PrimitiveDoubleConverter converter;
 	private @Mock MutableRequest request;
 	private @Mock HttpSession session;
 	private @Mock ServletContext context;
@@ -54,36 +53,36 @@ public class LocaleBasedFloatConverterTest {
 
 		when(request.getServletContext()).thenReturn(context);
 
-		converter = new LocaleBasedFloatConverter(new Locale("pt", "BR"));
+		converter = new PrimitiveDoubleConverter(new Locale("pt", "BR"));
 	}
 
 	@Test
 	public void shouldBeAbleToConvertWithPTBR() {
-		assertThat(converter.convert("10,00", Float.class), is(equalTo(new Float("10.00"))));
-		assertThat(converter.convert("10,01", Float.class), is(equalTo(new Float("10.01"))));
+		assertThat(converter.convert("10,00", double.class), is(equalTo(Double.parseDouble("10.00"))));
+		assertThat(converter.convert("10,01", double.class), is(equalTo(Double.parseDouble("10.01"))));
 	}
 
 	@Test
 	public void shouldBeAbleToConvertWithENUS() {
-		converter = new LocaleBasedFloatConverter(new Locale("en", "US"));
-		assertThat(converter.convert("10.00", Float.class), is(equalTo(new Float("10.00"))));
-		assertThat(converter.convert("10.01", Float.class), is(equalTo(new Float("10.01"))));
+		converter = new PrimitiveDoubleConverter(new Locale("en", "US"));
+		assertThat(converter.convert("10.00", double.class), is(equalTo(Double.parseDouble("10.00"))));
+		assertThat(converter.convert("10.01", double.class), is(equalTo(Double.parseDouble("10.01"))));
 	}
 
 	 @Test
 	 public void shouldBeAbleToConvertEmpty() {
-		 assertThat(converter.convert("", Float.class), is(nullValue()));
+		 assertThat(converter.convert("", double.class), is(equalTo(0d)));
 	 }
 
 	 @Test
 	 public void shouldBeAbleToConvertNull() {
-		 assertThat(converter.convert(null, Float.class), is(nullValue()));
+		 assertThat(converter.convert(null, double.class), is(equalTo(0d)));
 	 }
 
 	@Test
 	public void shouldThrowExceptionWhenUnableToParse() {
 		try {
-			converter.convert("vr3.9", Float.class);
+			converter.convert("vr3.9", double.class);
 			fail("Should throw exception");
 		} catch (ConversionException e) {
 			assertThat(e.getValidationMessage(), hasMessage("vr3.9 is not a valid number."));
