+++ vraptor-core/src/test/java/br/com/caelum/vraptor/ioc/GenericContainerTest.java
@@ -51,12 +51,12 @@ import br.com.caelum.vraptor.converter.ByteConverter;
 import br.com.caelum.vraptor.converter.Converter;
 import br.com.caelum.vraptor.converter.EnumConverter;
 import br.com.caelum.vraptor.converter.IntegerConverter;
-import br.com.caelum.vraptor.converter.LocaleBasedCalendarConverter;
-import br.com.caelum.vraptor.converter.LocaleBasedDateConverter;
-import br.com.caelum.vraptor.converter.LocaleBasedDoubleConverter;
-import br.com.caelum.vraptor.converter.LocaleBasedFloatConverter;
-import br.com.caelum.vraptor.converter.LocaleBasedPrimitiveDoubleConverter;
-import br.com.caelum.vraptor.converter.LocaleBasedPrimitiveFloatConverter;
+import br.com.caelum.vraptor.converter.CalendarConverter;
+import br.com.caelum.vraptor.converter.DateConverter;
+import br.com.caelum.vraptor.converter.DoubleConverter;
+import br.com.caelum.vraptor.converter.FloatConverter;
+import br.com.caelum.vraptor.converter.PrimitiveDoubleConverter;
+import br.com.caelum.vraptor.converter.PrimitiveFloatConverter;
 import br.com.caelum.vraptor.converter.LongConverter;
 import br.com.caelum.vraptor.converter.PrimitiveBooleanConverter;
 import br.com.caelum.vraptor.converter.PrimitiveByteConverter;
@@ -196,18 +196,18 @@ public abstract class GenericContainerTest {
 						put(long.class, PrimitiveLongConverter.class);
 						put(short.class, PrimitiveShortConverter.class);
 						put(byte.class, PrimitiveByteConverter.class);
-						put(double.class, LocaleBasedPrimitiveDoubleConverter.class);
-						put(float.class, LocaleBasedPrimitiveFloatConverter.class);
+						put(double.class, PrimitiveDoubleConverter.class);
+						put(float.class, PrimitiveFloatConverter.class);
 						put(boolean.class, PrimitiveBooleanConverter.class);
 						put(Integer.class, IntegerConverter.class);
 						put(Long.class, LongConverter.class);
 						put(Short.class, ShortConverter.class);
 						put(Byte.class, ByteConverter.class);
-						put(Double.class, LocaleBasedDoubleConverter.class);
-						put(Float.class, LocaleBasedFloatConverter.class);
+						put(Double.class, DoubleConverter.class);
+						put(Float.class, FloatConverter.class);
 						put(Boolean.class, BooleanConverter.class);
-						put(Calendar.class, LocaleBasedCalendarConverter.class);
-						put(Date.class, LocaleBasedDateConverter.class);
+						put(Calendar.class, CalendarConverter.class);
+						put(Date.class, DateConverter.class);
 						put(Enum.class, EnumConverter.class);
 					}
 					private static final long serialVersionUID = 8559316558416038474L;
